Implement a basic bridge FDB by learning.
The bridge FDB records a MAC address and its destination port that
is bound to the bridge device. When the next time the device attempts
to send a packet via the bridge device, it doesn't have to try to
deliver all ports of the bridge device.

The FDB will never be garbage collected until the nic device is removed
from the bridge.

PiperOrigin-RevId: 651533474